Reorganizes receivers
