+++ spec/Receiver/SeparateBusReceiverSpec.php
@@ -1,14 +1,12 @@
 <?php
 
-namespace spec\League\Tactician\Bernard;
+namespace spec\League\Tactician\Bernard\Receiver;
 
 use League\Tactician\CommandBus;
 use League\Tactician\Bernard\QueueableCommand;
-use League\Tactician\Bernard\QueuedCommand;
 use PhpSpec\ObjectBehavior;
-use Prophecy\Argument;
 
-class PreventLoopReceiverSpec extends ObjectBehavior
+class SeparateBusReceiverSpec extends ObjectBehavior
 {
     function let(CommandBus $commandBus)
     {
@@ -17,14 +15,21 @@ class PreventLoopReceiverSpec extends ObjectBehavior
 
     function it_is_initializable()
     {
-        $this->shouldHaveType('League\Tactician\Bernard\PreventLoopReceiver');
+        $this->shouldHaveType('League\Tactician\Bernard\Receiver\SeparateBusReceiver');
         $this->shouldHaveType('League\Tactician\Bernard\Receiver');
     }
 
     function it_handles_a_message(CommandBus $commandBus, QueueableCommand $command)
     {
-        $commandBus->handle(Argument::type('League\Tactician\Bernard\QueuedCommand'))->willReturn(true);
+        $commandBus->handle($command)->willReturn(true);
 
         $this->handle($command)->shouldReturn(true);
+    }
+
+    function it_is_invokable(CommandBus $commandBus, QueueableCommand $command)
+    {
+        $commandBus->handle($command)->willReturn(true);
+
+        $this->__invoke($command)->shouldReturn(true);
     }
 }
